+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+* List componentx:
+    * Add `minAutoColumnWidth` property (minimum column width, if width isn't defined in `userSettings`).
 
 ## [0.9.0-beta.16] - 2017-08-11
 ### Added